Move test to CacheTaskArchiveErrorIntegrationTest
